moved modules into lib directory as they are not external to project
