+++ couchapp/lib/couch.d.ts
