+++ couchapp/lib/couchapp.d.ts
@@ -1,4 +1,4 @@
-import {DesignDoc} from "couch";
+import {DesignDoc} from "./couch";
 
 export function loadAttachments(doc: DesignDoc, root: string, prefix?: string);
 export function loadFiles(dir:string, options?:{});
