+++ couchapp/lib/path.d.ts
